# openapi.yaml
openapi: 3.0.0
info:
  title: Plataforma de Streaming de Música API
  version: 1.0.0
  description: Uma API para gerenciar músicas, playlists, artistas e usuários.
paths:
  /usuarios:
    get:
      summary: Lista todos os usuários
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              example: { "usuarios": [...] }

    post:
      summary: Cria um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            example: { "nome_usuario": "novousuario", "email": "novousuario@email.com" }
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              example: { "mensagem": "Usuário criado com sucesso", "usuario": {...} }
  /usuarios/{usuarioId}:
    get:
      summary: Obtém detalhes de um usuário específico
      parameters:
        - name: usuarioId
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              example: { "usuario": {...} }

    put:
      summary: Atualiza um usuário existente
      parameters:
        - name: usuarioId
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example: { "nome_usuario": "novonome" }
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              example: { "mensagem": "Usuário atualizado com sucesso", "usuario": {...} }

    delete:
      summary: Exclui um usuário
      parameters:
        - name: usuarioId
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
      responses:
        '204':
          description: Usuário excluído com sucesso
...